refactor: move ticket animation to TicketAnimation.js
